Read the request body (Reverts #138)
I discovered, that when using Flask Swagger-UI and uploading 100 MByte files via formdata fields, not clearing the TCP receive buffer leads to a "Failed to Fetch" when using the browser. By emptying the TCP receive buffer, like it was done prior to version 4.5.0, the actual response is transported back to the user and no failed to fetch error occurs.